[WIP] Main code is moved to internal package
