+++ cmd/grcbot/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sattellite/tg-group-control-bot/internal/grcbot"
+
+func main() {
+	app := grcbot.Init()
+}
